updated Audit Record interface, added support for labeling with custom attack information

+++ label/utils.go
@@ -36,6 +36,8 @@ var (
 	excluded          = make(map[string]bool)
 
 	Debug bool
+
+	RemoveFilesWithoutMatches = false
 )
 
 func debug(args ...interface{}) {
@@ -84,11 +86,14 @@ func finish(wg *sync.WaitGroup, r *netcap.Reader, f *os.File, labelsTotal int, o
 		log.Fatal("failed to close", outFileName, ", error:", err)
 	}
 
-	// remove file that did not have any matching labels
-	if labelsTotal == 0 {
-		if err := os.Remove(outFileName); err != nil {
-			log.Fatal("failed remove empty file", outFileName, ", error:", err)
+	if RemoveFilesWithoutMatches {
+		// remove file that did not have any matching labels
+		if labelsTotal == 0 {
+			if err := os.Remove(outFileName); err != nil {
+				log.Fatal("failed remove empty file", outFileName, ", error:", err)
+			}
 		}
 	}
+
 	wg.Done()
 }
